+++ src/salesforce/data/dashboard.ts
@@ -7,7 +7,7 @@ import { Dashboard } from "..";
  * Queries Salesforce to get the Dashboard metadata using the standard API.
  * Caches the record ID if caching is enabled.
  */
-export async function getMetadataInfo(
+export async function retrieveRecord(
   metadataName: string,
   metadataType: FileType,
   context: vscode.ExtensionContext